chore(replicache): Add support for React Native using Expo SQLite
This is a minimal change to Replicache to ensure less disruptions.

To use you have to get the `expoSQLiteStoreProvider` from the `replicache/expo` package.

Usage:

```js
import {Replicache} from 'replicache';
import {expoSQLiteStoreProvider} from 'replicache/expo';

const rep = new Replicache({
  name: 'user',
  kvStore: expoSQLiteStoreProvider(),
});
```
