+++ packages/replicache/src/expo.ts
@@ -1,0 +1,2 @@
+export {expoSQLiteStoreProvider} from './expo/store.ts';
+export type {SQLiteDatabaseManagerOptions} from './kv/sqlite-store.ts';
