+++ packages/replicache/src/expo/store.ts
@@ -8,11 +8,10 @@ import {
   SQLiteDatabaseManager,
   type SQLiteDatabase,
   type SQLiteDatabaseManagerOptions,
-} from '../../replicache/src/kv/sqlite-store.ts';
-import type {StoreProvider} from '../../replicache/src/kv/store.ts';
-import {promiseVoid} from '../../shared/src/resolved-promises.ts';
+} from '../kv/sqlite-store.ts';
+import type {StoreProvider} from '../kv/store.ts';
 
-const expoDbManagerInstance = new SQLiteDatabaseManager({
+export const expoDbManagerInstance = new SQLiteDatabaseManager({
   open: fileName => {
     const db = openDatabaseSync(fileName);
 
@@ -41,24 +40,26 @@ const expoDbManagerInstance = new SQLiteDatabaseManager({
   },
 });
 
-export const expoSQLiteStoreProvider = (
+export function expoSQLiteStoreProvider(
   opts?: Partial<Omit<SQLiteDatabaseManagerOptions, 'journalMode'>>,
-): StoreProvider => ({
-  create: (name: string) =>
-    createSQLiteStore(expoDbManagerInstance)(name, {
-      // we default to 3 read connections for mobile devices
-      readPoolSize: 3,
-      busyTimeout: 200,
-      synchronous: 'NORMAL',
-      readUncommitted: false,
-      ...opts,
-      // we override the journal mode to undefined because
-      // setting it to WAL causes hanging COMMITs on Expo
-      journalMode: undefined,
-    }),
-  drop: (name: string) => {
-    expoDbManagerInstance.destroy(name);
+): StoreProvider {
+  return {
+    create: (name: string) =>
+      createSQLiteStore(expoDbManagerInstance)(name, {
+        // we default to 3 read connections for mobile devices
+        readPoolSize: 3,
+        busyTimeout: 200,
+        synchronous: 'NORMAL',
+        readUncommitted: false,
+        ...opts,
+        // we override the journal mode to undefined because
+        // setting it to WAL causes hanging COMMITs on Expo
+        journalMode: undefined,
+      }),
+    drop: (name: string) => {
+      expoDbManagerInstance.destroy(name);
 
-    return promiseVoid;
-  },
-});
+      return Promise.resolve();
+    },
+  };
+}
