+++ packages/replicache/src/http-status-unauthorized.ts
@@ -1,0 +1,1 @@
+export const httpStatusUnauthorized = 401;
