+++ packages/replicache/src/mod.ts
@@ -37,6 +37,7 @@ export type {
   StoreProvider as KVStoreProvider,
   Write as KVWrite,
 } from './kv/store.ts';
+export {makeIDBName} from './make-idb-name.ts';
 export {mergeAsyncIterables} from './merge-async-iterables.ts';
 export type {PatchOperation} from './patch-operation.ts';
 export type {PendingMutation} from './pending-mutations.ts';
@@ -49,7 +50,7 @@ export {
 export type {Puller, PullerResult, PullResponse} from './puller.ts';
 export type {Pusher, PusherResult, PushError, PushResponse} from './pusher.ts';
 export type {ReplicacheOptions} from './replicache-options.ts';
-export {makeIDBName, Replicache} from './replicache.ts';
+export {Replicache} from './replicache.ts';
 export {makeScanResult} from './scan-iterator.ts';
 export type {
   AsyncIterableIteratorToArray,
