+++ packages/replicache/tool/build.ts
@@ -31,7 +31,12 @@ async function buildReplicache(options: BuildOptions) {
   const outfile = basePath('out', 'replicache.' + ext);
   const entryPoints = {
     replicache: basePath('src', 'mod.ts'),
-    ...(forBundleSizeDashboard ? {} : {impl: basePath('src', 'impl.ts')}),
+    ...(forBundleSizeDashboard
+      ? {}
+      : {
+          impl: basePath('src', 'impl.ts'),
+          expo: basePath('src', 'expo.ts'),
+        }),
   };
   const result = await esbuild.build({
     ...sharedOptions(options.minify, metafile),
@@ -62,11 +67,12 @@ async function buildCLI() {
   });
 }
 
+const external = ['node:*', 'expo*'];
+
 if (perf) {
-  await buildReplicache({minify: true, ext: 'js', mode: 'release'});
+  await buildReplicache({minify: true, ext: 'js', mode: 'release', external});
 } else if (forBundleSizeDashboard) {
   // Bundle external modules for the bundle size dashboard
-  const external = ['node:*'];
   // We keep mjs as mjs so the dashboard does not get reset
   await Promise.all([
     buildReplicache({minify: false, ext: 'mjs', mode: 'unknown', external}),
@@ -75,5 +81,5 @@ if (perf) {
   ]);
 } else {
   await buildCLI();
-  await buildReplicache({minify: false, ext: 'js', mode: 'unknown'});
+  await buildReplicache({minify: false, ext: 'js', mode: 'unknown', external});
 }
