+++ packages/replicache/tsconfig.build.json
@@ -5,5 +5,5 @@
     "emitDeclarationOnly": true,
     "outDir": "out"
   },
-  "include": ["src/mod.ts", "src/impl.ts"]
+  "include": ["src/mod.ts", "src/impl.ts", "src/expo.ts"]
 }
