+++ packages/zero-react-native/src/mod.ts
@@ -1,1 +1,2 @@
-export {expoSQLiteStoreProvider} from './store.ts';
+export {expoSQLiteStoreProvider} from '../../replicache/src/expo/store.ts';
+export type {SQLiteDatabaseManagerOptions} from '../../replicache/src/kv/sqlite-store.ts';
